CARAMEL/requirements.txt
openai
streamlit
pypdf
xhtml2pdf
python-dotenv
python-docx
docx2pdf
